Add, active, or remove a card within card_tabset

Add, active, or remove a card within card_tabset

card_tabset_insert(
  inputId,
  title,
  ...,
  active = TRUE,
  notify_on_failure = TRUE,
  session = shiny::getDefaultReactiveDomain()
)

card_tabset_remove(
  inputId,
  title,
  notify_on_failure = TRUE,
  session = shiny::getDefaultReactiveDomain()
)

card_tabset_activate(
  inputId,
  title,
  notify_on_failure = TRUE,
  session = shiny::getDefaultReactiveDomain()
)

Arguments

inputId
the element id of card_tabset

title
the title of the card to insert, activate, or to remove

...
the content of the card

active
whether to set the card to be active once added

notify_on_failure
whether to show notifications on failure

session
shiny session domain

See also