Generates an 'accordion' tab-set

Generates an 'accordion' tab-set that only one tab is expanded at a time. This feature is experimental and has bugs in some situations. Please use it at your own risk.

accordion(
  ...,
  id = rand_string(),
  class = NULL,
  style_header = NULL,
  style_body = NULL,
  env = parent.frame(),
  extras = list(),
  root_path = template_root()
)

Arguments

...

'accordion' items, generated by accordion_item

id

the element id, must be unique

class

the additional 'HTML' class

style_header

additional 'CSS' styles for header

style_body

additional 'CSS' styles for content body

env

environment to evaluate ...

extras

key-value pairs that overrides the parameters in accordion_item

root_path

see template_root

Value

'shiny.tag.list' 'HTML' tags

See also

Examples

library(shiny)
library(shidashi)

accordion(
  id = "input-set",
  accordion_item(
    title = "Input Group A",
    textInput("input_1", "Input 1"),
    collapsed = FALSE,
    footer = "Anim pariatur cliche reprehenderit dolor brunch.",
    tools = list(
      as_badge("New|badge-danger")
      # card_tool(widget = "collapse")
    )
  ),
  accordion_item(
    title = "Input Group B",
    textInput("input_2", "Input 2"),
    footer = actionButton("btn1", "OK"),
    collapsed = FALSE,
    tools = list(
      card_tool(widget = "link",
                icon = shiny::icon("question-circle"),
                href = "#")
    )
  )
)
#> This Font Awesome icon ('question-circle') does not exist:
#> * if providing a custom `html_dependency` these `name` checks can 
#>   be deactivated with `verify_fa = FALSE`
#> <div id="input-set"><div class="card card-accordion ">
#>   <div class="card-header card-accordion" style="">
#>     <div class="card-tools card-accordion">
#>     <span class="right badge badge-danger">New</span>
#>   </div>
#>     <h4 class="card-title w-100 " data-toggle="collapse" href="#6TRMVCfoBI" aria-expanded="false">
#>       <a class="d-block w-100 card-accordion-title">
#>         Input Group A
#>       </a>
#>     </h4>
#>   </div>
#>   <div id="6TRMVCfoBI" class="collapse" data-parent="#input-set">
#>     <div class="card-body", style="">
#>       <div class="form-group shiny-input-container">
#>     <label class="control-label" id="input_1-label" for="input_1">Input 1</label>
#>     <input id="input_1" type="text" class="form-control" value=""/>
#>   </div>
#>     </div>
#>     <div class="card-footer">Anim pariatur cliche reprehenderit dolor brunch.</div>
#>   </div>
#> </div>
#> <div class="card card-accordion ">
#>   <div class="card-header card-accordion" style="">
#>     <div class="card-tools card-accordion">
#>     <a type="button" href="#" target="_self" class="btn btn-tool" data-card-widget="link">
#>       <i class="fa-question-circle fas" role="presentation" aria-label="question-circle icon"></i>
#>     </a>
#>   </div>
#>     <h4 class="card-title w-100 " data-toggle="collapse" href="#dvmZFaofL4" aria-expanded="false">
#>       <a class="d-block w-100 card-accordion-title">
#>         Input Group B
#>       </a>
#>     </h4>
#>   </div>
#>   <div id="dvmZFaofL4" class="collapse" data-parent="#input-set">
#>     <div class="card-body", style="">
#>       <div class="form-group shiny-input-container">
#>     <label class="control-label" id="input_2-label" for="input_2">Input 2</label>
#>     <input id="input_2" type="text" class="form-control" value=""/>
#>   </div>
#>     </div>
#>     <div class="card-footer">
#>     <button id="btn1" type="button" class="btn btn-default action-button">OK</button>
#>   </div>
#>   </div>
#> </div>
#> </div>