Obtain the module information

Obtain the module information

module_info(root_path = template_root(), settings_file = "modules.yaml")

load_module(
  root_path = template_root(),
  request = list(QUERY_STRING = "/"),
  env = parent.frame()
)

Arguments

root_path
the root path of the website project

settings_file
the settings file containing the module information

request
'HTTP' request string

env
environment to load module variables into

Details

The module files are stored in modules/ folder in your project. The folder names are the module id. Within each folder, there should be one "server.R", R/, and a "module-ui.html".

The R/ folder stores R code files that generate variables, which will be available to the other two files. These variables, along with some built-ins, will be used to render "module-ui.html". The built-in functions are

ns

shiny name-space function; should be used to generate the id for inputs and outputs. This strategy avoids conflict id effectively.

.module_id

a variable of the module id

module_title

a function that returns the module label

The "server.R" has access to all the code in R/ as well. Therefore it is highly recommended that you write each 'UI' component side-by-side with their corresponding server functions and call these server functions in "server.R".

Examples

library(shiny)
module_info()
#>        id order     group     label    icon      badge              url
#> 1    demo     0 Show Case Dashboard  circle New|bg-red    /?module=demo
#> 2    card     1      <NA>     Cards columns               /?module=card
#> 3 widgets     2      <NA>   Widgets      th            /?module=widgets

# load master module
load_module()
#> $environment
#> <environment: 0x55c3f3253040>
#> 
#> $has_module
#> [1] FALSE
#> 
#> $root_path
#> [1] "/home/runner/.local/share/R/shidashi/template"
#> 
#> $template_path
#> [1] "/home/runner/.local/share/R/shidashi/template/index.html"
#> 
#> $module
#> $module$id
#> NULL
#> 
#> $module$server
#> function (input, output, session, ...) 
#> {
#> }
#> <bytecode: 0x55c3f4681e80>
#> <environment: 0x55c3f4684190>
#> 
#> $module$template_path
#> NULL
#> 
#> 

# load specific module
module_data <- load_module(
  request = list(QUERY_STRING = "/?module=card"))
env <- module_data$environment

# get module title
env$module_title()
#> [1] "Cards"

# generate module-specific shiny id
env$ns("input1")
#> [1] "card-input1"

# generate part of the UI
env$ui_card2()
#> <div class="col-sm-4"><div class="card card-single direct-chat  ">
#>   <div class="card-header " style="">
#>     <h3 class="card-title">Card2 Example</h3>
#>     <!-- Buttons, labels, and many other things can be placed here! -->
#>     <div class="card-tools">
#>       <a type="button" href="https://github.com/dipterix" target="_blank" class="btn btn-tool" data-card-widget="link">
#>     <i class="fa-external-link-alt fas" role="presentation" aria-label="external-link-alt icon"></i>
#>   </a>
#>   <a type="button" href="#" target="_self" class="btn btn-tool" data-card-widget="refresh">
#>     <i class="fa-sync-alt fas" role="presentation" aria-label="sync-alt icon"></i>
#>   </a>
#>   <a type="button" href="#" target="_self" class="btn btn-tool" data-card-widget="collapse">
#>     <i class="fa-minus fas" role="presentation" aria-label="minus icon"></i>
#>   </a>
#>   <a type="button" href="#" target="_self" class="btn btn-tool" data-card-widget="maximize">
#>     <i class="fa-expand fas" role="presentation" aria-label="expand icon"></i>
#>   </a>
#>       <button type="button" class="btn btn-tool card2-switch" title="Configurations" data-widget="chat-pane-toggle">
#>         <i class="fas fa-ellipsis-v"></i>
#>       </button>
#>     </div>
#>   </div>
#>   <div class="card-body card2-body height-300" style=";">
#>     <div class="direct-chat-messages fill no-padding overflow-hidden">
#>       <div id="card-card2_plot" class="shiny-plot-output" style="width:100%;height:100%;"></div>
#>     </div>
#>     <div class="direct-chat-contacts fill padding-20 opacity-07">
#>       <div class="row">
#>     <div class="col-sm-6">
#>       <div class="form-group shiny-input-container">
#>         <label class="control-label" id="card-card2_plot_title-label" for="card-card2_plot_title">Plot title</label>
#>         <input id="card-card2_plot_title" type="text" class="form-control" value=""/>
#>       </div>
#>     </div>
#>     <div class="col-sm-6">
#>       <div class="form-group shiny-input-container">
#>         <label class="control-label" id="card-card2_plot_npts-label" for="card-card2_plot_npts"># of points</label>
#>         <input class="js-range-slider" id="card-card2_plot_npts" data-skin="shiny" data-min="1" data-max="100" data-from="10" data-step="1" data-grid="true" data-grid-num="9.9" data-grid-snap="false" data-prettify-separator="," data-prettify-enabled="true" data-keyboard="true" data-data-type="number"/>
#>       </div>
#>     </div>
#>   </div>
#>     </div>
#>   </div>
#>   
#> </div>
#> </div>
#> <div class="col-sm-8">
#>   <div class="card card-tabs ">
#>     <div class="card-header p-0 pt-1 ">
#>       <ul class="nav nav-tabs" id="card-card2_code_cardset" role="tablist">
#>         <li class="nav-item nav-tab-header">
#>           <a class="nav-link active" id="card-card2_code_cardset-1-tab" data-toggle="pill" href="#card-card2_code_cardset-1" role="tab" aria-controls="card-card2_code_cardset-1" aria-selected="true" tab-index="1">UI code</a>
#>         </li>
#>         <li class="nav-item nav-tab-header">
#>           <a class="nav-link" id="card-card2_code_cardset-2-tab" data-toggle="pill" href="#card-card2_code_cardset-2" role="tab" aria-controls="card-card2_code_cardset-2" aria-selected="false" tab-index="2">server.R</a>
#>         </li>
#>       </ul>
#>     </div>
#>     <div class="card-body height-300 overflow-y-auto">
#>       <div class="tab-content" id="card-card2_code_cardsetContent">
#>         <div class="tab-pane fade active fill-min-height show position-relative" id="card-card2_code_cardset-1" role="tabpanel" aria-labelledby="card-card2_code_cardset-1-tab" tab-index="1">
#>           <div class="row">
#>             <div class="col-sm-12"><pre class='padding-8 no-margin bg-gray-90 overflow-auto  clipboard-btn shidashi-clipboard-output' data-clipboard-text='card2(
#>   title = "Card2 Example", body_main = plotOutput(
#>     outputId = ns("card2_plot"),
#>     height = "100%"
#>   ),
#>   body_side = fluidRow(
#>     column(
#>       6L, textInput(
#>         ns("card2_plot_title"),
#>         "Plot title"
#>       )
#>     ),
#>     column(
#>       6L, sliderInput(
#>         ns("card2_plot_npts"),
#>         "# of points", min = 1, max = 100, value = 10, step = 1, round = TRUE
#>       )
#>     )
#>   ),
#>   tools = list(
#>     card_tool(widget = "link", href = "https://github.com/dipterix"),
#>     card_tool(widget = "refresh"),
#>     card_tool(widget = "collapse"),
#>     card_tool(widget = "maximize")
#>   ),
#>   class_body = "height-300"
#> )' role='button' title='Click to copy!'><code class='r'>card2(
#>   title = "Card2 Example", body_main = plotOutput(
#>     outputId = ns("card2_plot"),
#>     height = "100%"
#>   ),
#>   body_side = fluidRow(
#>     column(
#>       6L, textInput(
#>         ns("card2_plot_title"),
#>         "Plot title"
#>       )
#>     ),
#>     column(
#>       6L, sliderInput(
#>         ns("card2_plot_npts"),
#>         "# of points", min = 1, max = 100, value = 10, step = 1, round = TRUE
#>       )
#>     )
#>   ),
#>   tools = list(
#>     card_tool(widget = "link", href = "https://github.com/dipterix"),
#>     card_tool(widget = "refresh"),
#>     card_tool(widget = "collapse"),
#>     card_tool(widget = "maximize")
#>   ),
#>   class_body = "height-300"
#> )</code></pre></div>
#>           </div>
#>         </div>
#>         <div class="tab-pane fade fill-min-height position-relative" id="card-card2_code_cardset-2" role="tabpanel" aria-labelledby="card-card2_code_cardset-2-tab" tab-index="2">
#>           <div class="row">
#>             <div class="col-sm-12"><pre class='padding-8 no-margin bg-gray-90 overflow-auto  clipboard-btn shidashi-clipboard-output' data-clipboard-text='output$card2_plot <- renderPlot(
#>   {
#>     npoints <- input$card2_plot_npts
#>     title <- input$card2_plot_title
#>     if (!length(title) ||
#>       title == "") {
#>       title <- "Normal Q-Q Plot"
#>     }
#>     qqnorm(
#>       rnorm(npoints),
#>       main = title
#>     )
#>     abline(a = 0, b = 1, col = "orange3", lty = 2)
#>   }
#> )' role='button' title='Click to copy!'><code class='r'>output$card2_plot <- renderPlot(
#>   {
#>     npoints <- input$card2_plot_npts
#>     title <- input$card2_plot_title
#>     if (!length(title) ||
#>       title == "") {
#>       title <- "Normal Q-Q Plot"
#>     }
#>     qqnorm(
#>       rnorm(npoints),
#>       main = title
#>     )
#>     abline(a = 0, b = 1, col = "orange3", lty = 2)
#>   }
#> )</code></pre></div>
#>           </div>
#>         </div>
#>       </div>
#>     </div>
#>   </div>
#> </div>