Obtain the module information

Obtain the module information

module_info(root_path = template_root(), settings_file = "modules.yaml")

load_module(
  root_path = template_root(),
  request = list(QUERY_STRING = "/"),
  env = parent.frame()
)

Arguments

root_path
the root path of the website project

settings_file
the settings file containing the module information

request
'HTTP' request string

env
environment to load module variables into

Details

The module files are stored in modules/ folder in your project. The folder names are the module id. Within each folder, there should be one "server.R", R/, and a "module-ui.html".

The R/ folder stores R code files that generate variables, which will be available to the other two files. These variables, along with some built-ins, will be used to render "module-ui.html". The built-in functions are

ns

shiny name-space function; should be used to generate the id for inputs and outputs. This strategy avoids conflict id effectively.

.module_id

a variable of the module id

module_title

a function that returns the module label

The "server.R" has access to all the code in R/ as well. Therefore it is highly recommended that you write each 'UI' component side-by-side with their corresponding server functions and call these server functions in "server.R".

Examples

library(shiny)
module_info()
#>          id group        label   icon      badge                url
#> 1 module_id    NA Module Label circle New|bg-red /?module=module_id

# load master module
load_module()
#> $environment
#> <environment: 0x55dbed3e55f0>
#> 
#> $has_module
#> [1] FALSE
#> 
#> $root_path
#> [1] "/home/runner/.local/share/R/shidashi/AdminLTE3-bare"
#> 
#> $template_path
#> [1] "/home/runner/.local/share/R/shidashi/AdminLTE3-bare/index.html"
#> 
#> $module
#> $module$id
#> NULL
#> 
#> $module$server
#> function (input, output, session, ...) 
#> {
#> }
#> <bytecode: 0x55dbece95e98>
#> <environment: 0x55dbece981a8>
#> 
#> $module$template_path
#> NULL
#> 
#> 

# load specific module
module_data <- load_module(
  request = list(QUERY_STRING = "/?module=module_id"))
env <- module_data$environment

# get module title
env$module_title()
#> [1] "Module Label"

# generate module-specific shiny id
env$ns("input1")
#> [1] "module_id-input1"

# generate part of the UI
env$ui()
#> <div class="container-fluid">
#>   <div class="row">
#>     <div class="col-sm-12">
#>       <div id="module_id-plot" class="shiny-plot-output" style="width:100%;height:400px;"></div>
#>     </div>
#>   </div>
#> </div>